backend/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone-client
langchain
langchain-openai
python-multipart
pypdf
langchain_community
langchain_pinecone
langchain-classic
python-docx